+++ data/src/main/java/com/zaroslikov/data/room/dao/AnimalSizeDao.kt
@@ -25,7 +25,7 @@ interface AnimalSizeDao {
                 " WHERE idAnimal=:id" +
                 " ORDER BY DATE(printf('%04d-%02d-%02d', substr(date, 7, 4), substr(date, 4, 2), substr(date, 1, 2))) DESC, id DESC"
     )
-    fun getSizeAnimalLimit(id: Int): Flow<AnimalSizeTable>
+    fun getSizeAnimalLimit(id: Long): Flow<AnimalSizeTable>
 
     @Query(
         "SELECT id, size, suffix, date, idAnimal, note" +
@@ -33,5 +33,5 @@ interface AnimalSizeDao {
                 " WHERE idAnimal=:id" +
                 " ORDER BY DATE(printf('%04d-%02d-%02d', substr(date, 7, 4), substr(date, 4, 2), substr(date, 1, 2))) DESC, id DESC"
     )
-    fun getSizeAnimal(id: Int): Flow<List<AnimalSizeTable>>
+    fun getSizeAnimal(id: Long): Flow<List<AnimalSizeTable>>
 }
