+++ data/src/main/java/com/zaroslikov/data/room/dao/AnimalWeightDao.kt
@@ -24,12 +24,12 @@ interface AnimalWeightDao {
                 " WHERE idAnimal=:id" +
                 " ORDER BY DATE(printf('%04d-%02d-%02d', substr(date, 7, 4), substr(date, 4, 2), substr(date, 1, 2))) DESC, id DESC"
     )
-    fun getWeightAnimalLimit(id: Int): Flow<AnimalWeightTable>
+    fun getWeightAnimalLimit(id: Long): Flow<AnimalWeightTable>
 
     @Query(
         "SELECT  * FROM AnimalWeightTable" +
                 " WHERE idAnimal=:id" +
                 " ORDER BY DATE(printf('%04d-%02d-%02d', substr(date, 7, 4), substr(date, 4, 2), substr(date, 1, 2))) DESC, id DESC"
     )
-    fun getWeightAnimal(id: Int): Flow<List<AnimalWeightTable>>
+    fun getWeightAnimal(id: Long): Flow<List<AnimalWeightTable>>
 }
