+++ domain/src/main/java/com/zaroslikov/domain/repository/AnimalSizeRepository.kt
@@ -8,6 +8,6 @@ interface AnimalSizeRepository {
     suspend fun insertAnimalSizeTable(animalSizeTable: DomainAnimalSize)
     suspend fun updateAnimalSizeTable(animalSizeTable: DomainAnimalSize)
     suspend fun deleteAnimalSizeTable(animalSizeTable: DomainAnimalSize)
-    fun getSizeAnimalLimit(id: Int): Flow<DomainAnimalSize>
-    fun getSizeAnimal(id: Int): Flow<List<DomainAnimalSize>>
+    fun getSizeAnimalLimit(id: Long): Flow<DomainAnimalSize>
+    fun getSizeAnimal(id: Long): Flow<List<DomainAnimalSize>>
 }
